Move professor folder
